// src/interfaces/ListQueue.ts
export interface ListQueue {
  queue_er_id: number
  hn: string
  vn: string
  queue_number: string
  urgency_type_id: number
  service_point_id: number
  date_serv: string
  time_serv: string
  queue_status_id: any
  date_create: string
  date_update: string
  is_completed: string
  is_cancel: string
  title: string
  first_name: string
  last_name: string
  birthdate: string
  remark: any
  update_date: string
  sex: string
  urgency_type_name: string
  urgency_type_name_th: string
  urgency_type_color: string
  urgency_type_order: string
}